rename src/cadl-generated to src/generated
