+++ sdk/eventgrid/eventgrid-namespaces/src/models.ts
@@ -11,7 +11,7 @@ import type {
   RenewCloudEventLocksOptionalParams,
   EventGridClientOptions as EventGridOptions,
   ReleaseDelay,
-} from "./cadl-generated/index.js";
+} from "./generated/index.js";
 
 /** Send Event Options */
 export interface SendEventOptions extends OperationOptions {
